Move connection setup into a separate function.
